tidy up diffusers branch backporting of cross attention refactoring
