+++ ldm/models/diffusion/shared_invokeai_diffusion.py
@@ -5,9 +5,8 @@ from typing import Callable, Optional, Union
 import torch
 
 from ldm.models.diffusion.cross_attention_control import Arguments, \
-    remove_cross_attention_control, setup_cross_attention_control, Context
-from ldm.models.diffusion.cross_attention_map_saving import AttentionMapSaver, setup_attention_map_saving, \
-    remove_attention_map_saving
+    remove_cross_attention_control, setup_cross_attention_control, Context, get_cross_attention_modules, CrossAttentionType
+from ldm.models.diffusion.cross_attention_map_saving import AttentionMapSaver
 
 
 class InvokeAIDiffuserComponent:
@@ -54,11 +53,24 @@ class InvokeAIDiffuserComponent:
         remove_cross_attention_control(self.model)
 
     def setup_attention_map_saving(self, saver: AttentionMapSaver):
-        setup_attention_map_saving(self.model, saver)
+        def callback(slice, dim, offset, slice_size, key):
+            if dim is not None:
+                print("sliced tokens attention map saving is not implemented")
+                return
+            saver.add_attention_maps(slice, key)
+
+        tokens_cross_attention_modules = get_cross_attention_modules(self.model, CrossAttentionType.TOKENS)
+        for identifier, module in tokens_cross_attention_modules:
+            key = ('down' if identifier.startswith('down') else
+                   'up' if identifier.startswith('up') else
+                   'mid')
+            module.set_attention_slice_calculated_callback(
+                lambda slice, dim, offset, slice_size, key=key: callback(slice, dim, offset, slice_size, key))
 
     def remove_attention_map_saving(self):
-        remove_attention_map_saving(self.model)
-
+        tokens_cross_attention_modules = get_cross_attention_modules(self.model, CrossAttentionType.TOKENS)
+        for _, module in tokens_cross_attention_modules:
+            module.set_attention_slice_calculated_callback(None)
 
     def do_diffusion_step(self, x: torch.Tensor, sigma: torch.Tensor,
                                 unconditioning: Union[torch.Tensor,dict],
